requirements.txt
nicegui
requests_oauthlib
python-dotenv
firebase-admin
google-auth-oauthlib
google-auth